demos/Insurance_Advisor_Chatbot_GitHub/agent/insurance_advisor/app_utils/.requirements.txt
a2a-sdk==0.3.26
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1 ; python_full_version < '3.11'
asyncpg==0.31.0
attrs==26.1.0
authlib==1.7.0
bcrypt==5.0.0
build==1.4.4
certifi==2026.4.22
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6 ; os_name == 'nt' or sys_platform == 'win32'
cryptography==47.0.0
decorator==5.2.1
distro==1.9.0
docstring-parser==0.18.0
durationpy==0.10
exceptiongroup==1.3.1 ; python_full_version < '3.11'
fastapi==0.136.1
fastuuid==0.14.0
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
gcsfs==2026.3.0
google-adk==1.31.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.149.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-iamconnectorcredentials==0.1.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-storage-control==1.11.0
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.5.0 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.12.0
idna==3.13
importlib-metadata==8.5.0
importlib-resources==7.1.0
jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
joserfc==1.6.4
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
litellm==1.83.7
mako==1.3.11
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0 ; python_full_version < '3.11'
multidict==6.7.1
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.4.4 ; python_full_version >= '3.11'
oauthlib==3.3.1
onnxruntime==1.24.3 ; python_full_version < '3.11'
onnxruntime==1.25.1 ; python_full_version >= '3.11'
openai==2.30.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-google-genai==0.7b0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-genai==0.3b0
orjson==3.11.8
overrides==7.7.0
packaging==26.2
pandas==2.3.3 ; python_full_version < '3.11'
pandas==3.0.2 ; python_full_version >= '3.11'
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pybase64==1.4.3
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.14.0
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.1.0
pyparsing==3.3.2
pypika==0.51.1
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.27
pytz==2026.1.post1 ; python_full_version < '3.11'
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
ruamel-yaml==0.19.1
scikit-learn==1.5.2 ; python_full_version < '3.11'
scikit-learn==1.8.0 ; python_full_version >= '3.11'
scipy==1.15.3 ; python_full_version < '3.11'
scipy==1.17.1 ; python_full_version >= '3.11'
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.4.1
starlette==0.52.1
sympy==1.14.0 ; python_full_version < '3.11'
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.1 ; python_full_version < '3.11'
tqdm==4.67.3
typer==0.23.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.2 ; python_full_version < '3.11' or sys_platform == 'emscripten' or sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.34.3
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.1
